+++ src/game/player/libElements.ts
@@ -1,17 +1,13 @@
 import Isolated from "@player/lib/isolated";
 import Say from "@player/elements/say/Say";
-import {Top, Center, Bottom, Full} from "@player/lib/PlayerFrames";
-import {VBox, HBox} from "@player/lib/UtilComponents";
 import {usePreference} from "@player/lib/preferences";
+import {Stage} from "@player/lib/PageRouter/Stage";
+import {Page} from "@player/lib/PageRouter/Page";
 
 export {
     Isolated,
     Say,
-    Top,
-    Center,
-    Bottom,
-    VBox,
-    HBox,
-    Full,
     usePreference,
+    Stage,
+    Page,
 };